+++ inc/ORM/AbstractModel.php
@@ -13,16 +13,7 @@ abstract class AbstractModel {
     protected helper_plugin_sqlite $sqlite;
 
     public function __construct(helper_plugin_sqlite $sqlite) {
-
         $this->sqlite = $sqlite;
-        if (!$this->sqlite->getAdapter()->getDb()) {
-            // $this->sqlite->getAdapter()->opendb(false);
-        }
-    }
-
-    public function findMaxNewsId(): int {
-        $res = $this->sqlite->query('SELECT max(news_id) FROM news');
-        return (int)$this->sqlite->res2single($res);
     }
 
     abstract public static function createFromArray(\helper_plugin_sqlite $helperPluginSqlite, array $data): self;
