+++ inc/ORM/AbstractService.php
@@ -3,7 +3,7 @@
 namespace FYKOS\dokuwiki\Extension\PluginNewsFeed\ORM;
 
 use FYKOS\dokuwiki\Extension\PluginNewsFeed\Model\AbstractModel;
-use FYKOS\dokuwiki\Extension\PluginNewsFeed\Model\ModelStream;
+use helper_plugin_sqlite;
 
 /**
  * Class AbstractService
@@ -15,33 +15,35 @@ abstract class AbstractService {
     /** @var string|AbstractModel */
     protected string $modelClassName;
 
-    protected \helper_plugin_sqlite $sqlite;
+    protected helper_plugin_sqlite $sqlite;
 
-    public function __construct(\helper_plugin_sqlite $sqlite, string $table, string $modelClassName) {
+    public function __construct(helper_plugin_sqlite $sqlite, string $table, string $modelClassName) {
         $this->table = $table;
         $this->sqlite = $sqlite;
         $this->modelClassName = $modelClassName;
     }
 
-    public function getById(int $id): AbstractModel {
-        $res = $this->sqlite->query('SELECT * FROM ? WHERE ' . $this->table . '_id = ?', $this->table, (int)$id);
-        return ($this->modelClassName)::createFromArray($this->sqlite, $res->fetch());
+    public function getById(int $id): ?AbstractModel {
+        $res = $this->sqlite->query('SELECT * FROM ' . $this->table . ' WHERE ' . $this->table . '_id = ?', (int)$id)->fetch();
+        if ($res) {
+            return ($this->modelClassName)::createFromArray($this->sqlite, $res);
+        }
+        return null;
     }
 
     public function create(array $data): bool {
-        $sql = 'INSERT ? INTO (' . join(',', array_keys($data)) . ')  VALUES( ' . join(',', array_fill(0, count($data), '?')) . ')';
+        $sql = 'INSERT INTO ' . $this->table . ' (' . join(',', array_keys($data)) . ')  VALUES( ' . join(',', array_fill(0, count($data), '?')) . ')';
 
         return (bool)$this->sqlite->query($sql,
-            $this->table,
-            ...array_values($data));
+            ...array_values($data)
+        );
     }
 
-    public function update(AbstractModel $model, array $data) {
-        $sql = 'UPDATE ? SET ' . join(',', array_map(function ($key) {
+    public function update(AbstractModel $model, array $data): void {
+        $sql = 'UPDATE ' . $this->table . ' SET ' . join(',', array_map(function ($key) {
                 return $key . '=?';
             }, array_keys($data))) . ' WHERE ' . $this->table . '_id' . '=?';
         $this->sqlite->query($sql,
-            $this->table,
             ...array_values($data),
             ...[
                 $model->{$this->table . 'Id'},
@@ -50,7 +52,7 @@ abstract class AbstractService {
     }
 
     public function getMaxId(): int {
-        $res = $this->sqlite->query('SELECT max(?) FROM ?', $this->table . '_id', $this->table);
+        $res = $this->sqlite->query('SELECT max(' . $this->table . '_id) FROM ?', $this->table);
         return (int)$this->sqlite->res2single($res);
     }
